pyproject.toml
[project]
name = "cloud_run_mcp_proxy"
version = "0.0.0"
description = "Proxy to access remote MCP servers running on GCP cloud run."
authors = [
  {name="Jasper Ginn", email="jasperginn@gmail.com"}
]
requires-python = ">=3.12"
readme = "README.md"
dependencies = [
    "cachetools>=5.5.2",
    "fastmcp>=2.7.1",
    "google-auth>=2.40.3",
]

[project.scripts]
cloud_run_mcp_proxy = "cloud_run_mcp_proxy.proxy:entrypoint"

[tool.ruff]
lint.ignore = ["E501"]
extend-exclude = [
  "__pycache__",
  ".eggs",
  ".git",
  ".venv",
  "build",
  "dist",
  "notebooks",
  ".cache"
]
line-length = 100
indent-width = 4

[tool.ruff.format]
quote-style = "single"
indent-style = "space"

[tool.pyright]
exclude = [".venv", ".github", "docs", "tests", ".devcontainer"]
include = ["src", "scripts"]
pythonVersion = "3.12"
venvPath = "."
venv = ".venv"

[tool.pytest.ini_options]
markers = [
  "integration: marks integration tests",
  "llm: marks tests that require LLM calls",
]
cache_dir = "/home/vscode/workspace/.cache/pytest"

[dependency-groups]
dev = [
    "nest-asyncio>=1.6.0",
    "pre-commit>=3.8.0",
    "ipykernel>=6.29.5",
]
test = [
    "pytest-asyncio>=1.0.0",
    "aioresponses>=0.7.8",
    "pytest>=8.3.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
